refactor: 교차점 네이밍 변경에 따른 기타 네이밍 수정
