+++ src/main/kotlin/omok/domain/omokboard/Intersection.kt
@@ -1,16 +1,16 @@
 package omok.domain.omokboard
 
-import omok.domain.omokboard.PointState.EMPTY
-import omok.domain.omokboard.PointState.OCCUPIED_BLACK
-import omok.domain.omokboard.PointState.OCCUPIED_WHITE
+import omok.domain.omokboard.IntersectionState.EMPTY
+import omok.domain.omokboard.IntersectionState.OCCUPIED_BLACK
+import omok.domain.omokboard.IntersectionState.OCCUPIED_WHITE
 import omok.domain.player.StoneColor
 import omok.domain.player.StoneColor.BLACK
 import omok.domain.player.StoneColor.WHITE
 
 data class Intersection(
-    private var _state: PointState = EMPTY,
+    private var _state: IntersectionState = EMPTY,
 ) {
-    val state: PointState get() = _state
+    val state: IntersectionState get() = _state
 
     fun updateState(stoneColor: StoneColor) {
         _state =
