+++ src/main/kotlin/omok/domain/omokboard/IntersectionState.kt
@@ -1,6 +1,6 @@
 package omok.domain.omokboard
 
-enum class PointState {
+enum class IntersectionState {
     OCCUPIED_BLACK,
     OCCUPIED_WHITE,
     EMPTY,
