+++ src/main/kotlin/omok/domain/rule/DrawRule.kt
@@ -1,7 +1,7 @@
 package omok.domain.rule
 
+import omok.domain.omokboard.IntersectionState
 import omok.domain.omokboard.OmokBoard
-import omok.domain.omokboard.PointState
 import omok.domain.player.PlayerStone
 
 class DrawRule : OmokRule {
@@ -9,7 +9,7 @@ class DrawRule : OmokRule {
         omokBoard: OmokBoard,
         playerStone: PlayerStone,
     ): PlaceResult =
-        if (omokBoard.value.values.count { it.state == PointState.EMPTY } == 1) {
+        if (omokBoard.value.values.count { it.state == IntersectionState.EMPTY } == 1) {
             PlaceResult.Success.Finish(GameResult.DRAW)
         } else {
             PlaceResult.Success.Progress(playerStone)
